선정 평가 상세 기능을 구현한다.

+++ src/main/java/com/yujigyeongseong/api/domain/work_lounge/service/SelectEvaluationService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface SelectEvaluationService {
     public List<SelectEvaluationDTO> getSelectEvaluationList();
 
-    public List<SelectEvaluationDTO> getselectEvaluation(Integer subAnnNo);
+    public List<SelectEvaluationDTO> getselectEvaluation(Long subAnnNo);
 
     public List<SelectEvaluationDTO> getselectCommittee(String rndTaskNo);
 }
